Remove explicit sonar scan of PR.
I cannot for the life of me figure out how to get the sonar token to be
available in the PR.  Setting it in the project doesn't work for PRs
since tokens are not available to PRs run from forks.  Setting it in my
fork didn't work either.  It is available via codecov so /shrug